Fix Hibernate setup
Add mappings to Hibernate configuration file.

Add the missing one-to-one relation between DailyReport and
VendingMachine.

Change all IDs to Long and money variables to BigDecimal.

Change constructor of ProductReport class to avoid the null ID
problem. Now to create a instance of this class both Product and
DailyReport are needed.